# openapi/swagger.json
{
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Autonomous delivery-integrity workflow engine for LUME streaming operations.",
        "title": "Fincher Media Delivery Operations API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/titles": {
            "get": {
                "description": "Fetches releases in the launch calendar with pagination, optionally filtered by status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "titles"
                ],
                "summary": "List all media titles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status filter (ON_TRACK, AT_RISK, HOLD, PROCESSING, SHIPPED)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 10, max: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (asc, desc)",
                        "name": "sort_order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TitlePaginationResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Inserts a new release title into the launch calendar, bundles Master V01, and triggers holistic allocation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "titles"
                ],
                "summary": "Create a media title",
                "parameters": [
                    {
                        "description": "Title payload",
                        "name": "title",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Title"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Title"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.DomainError"
                        }
                    }
                }
            }
        },
        "/titles/{id}": {
            "get": {
                "description": "Fetches detailed launch metadata for a specific title.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "titles"
                ],
                "summary": "Get media title by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Title"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.DomainError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Removes a title if it has no dependent packages or masters.",
                "tags": [
                    "titles"
                ],
                "summary": "Delete a media title",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.DomainError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errors.DomainError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates specific fields (overall_status, master version, metadata, premiere_date) on a title. When premiere_date is updated, re-arms the deadline timer and re-drives the workflow.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "titles"
                ],
                "summary": "Partial update of a media title",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Partial title updates",
                        "name": "update",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateTitleInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Title"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.DomainError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.DomainError"
                        }
                    }
                }
            }
        },
        "/titles/{id}/qc": {
            "post": {
                "description": "Initiates master cut quality control inspection for a title. Rejects if the title is already in QC.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "titles"
                ],
                "summary": "Send Title for Master QC",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Title"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.DomainError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errors.DomainError": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/errors.ErrorCode"
                },
                "message": {
                    "type": "string"
                },
                "op": {
                    "type": "string"
                }
            }
        },
        "errors.ErrorCode": {
            "type": "string",
            "enum": [
                "NOT_FOUND",
                "ALREADY_EXISTS",
                "INVALID_INPUT",
                "CONFLICT",
                "INTERNAL",
                "BUDGET_EXCEEDED",
                "GATE_TERMINATED",
                "UNAUTHENTICATED",
                "UNAUTHORIZED"
            ],
            "x-enum-varnames": [
                "CodeNotFound",
                "CodeAlreadyExists",
                "CodeInvalidInput",
                "CodeConflict",
                "CodeInternal",
                "CodeBudgetExceeded",
                "CodeGateTerminated",
                "CodeUnauthenticated",
                "CodeUnauthorized"
            ]
        },
        "errors.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "op": {
                    "type": "string"
                }
            }
        },
...
        "models.Title": {
            "type": "object",
            "required": [
                "current_master_version",
                "id",
                "name",
                "overall_status",
                "premiere_date",
                "slug",
                "territories",
                "type"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "current_master_version": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "name": {
                    "type": "string"
                },
                "overall_status": {
                    "enum": [
                        "DRAFT",
                        "ON_TRACK",
                        "AT_RISK",
                        "HOLD",
                        "PROCESSING",
                        "SHIPPED",
                        "OVERDUE"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TitleStatus"
                        }
                    ]
                },
                "premiere_date": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "territories": {
                    "type": "integer",
                    "minimum": 1
                },
                "type": {
                    "enum": [
                        "FEATURE",
                        "SERIES",
                        "SPECIAL"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TitleType"
                        }
                    ]
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.TitlePaginationResult": {
            "type": "object",
            "properties": {
                "has_next_page": {
                    "type": "boolean"
                },
                "has_prev_page": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Title"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "models.TitleStatus": {
            "type": "string",
            "enum": [
                "DRAFT",
                "ON_TRACK",
                "AT_RISK",
                "HOLD",
                "PROCESSING",
                "SHIPPED",
                "OVERDUE"
            ],
            "x-enum-varnames": [
                "StatusDraft",
                "StatusOnTrack",
                "StatusAtRisk",
                "StatusHold",
                "StatusProcessing",
                "StatusShipped",
                "StatusOverdue"
            ]
        },
        "models.TitleType": {
            "type": "string",
            "enum": [
                "FEATURE",
                "SERIES",
                "SPECIAL"
            ],
            "x-enum-varnames": [
                "TitleTypeFeature",
                "TitleTypeSeries",
                "TitleTypeSpecial"
            ]
        },
...
        "models.UpdateTitleInput": {
            "type": "object",
            "properties": {
                "current_master_version": {
                    "type": "string",
                    "minLength": 1
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "name": {
                    "type": "string",
                    "minLength": 1
                },
                "overall_status": {
                    "enum": [
                        "DRAFT",
                        "ON_TRACK",
                        "AT_RISK",
                        "HOLD",
                        "PROCESSING",
                        "SHIPPED",
                        "OVERDUE"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TitleStatus"
                        }
                    ]
                },
                "premiere_date": {
                    "type": "string"
                },
                "slug": {
                    "type": "string",
                    "minLength": 1
                },
                "territories": {
                    "type": "integer",
                    "minimum": 1
                },
                "type": {
                    "enum": [
                        "FEATURE",
                        "SERIES",
                        "SPECIAL"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TitleType"
                        }
                    ]
                }
            }
        },
...
    }
}